Se cambio la forma de manejar las vistas y el layput
