+++ package.json
@@ -11,6 +11,8 @@
   },
   "dependencies": {
     "@ant-design/v5-patch-for-react-19": "^1.0.3",
+    "@stripe/react-stripe-js": "^3.5.1",
+    "@stripe/stripe-js": "^6.1.0",
     "antd": "^5.24.3",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
