+++ src/App.jsx
@@ -1,54 +1,35 @@
 //React
 import { Routes, Route } from 'react-router-dom';
-import { CollapsedProvider } from './Context/CollapseContext';
+
+//Layout
+import AdminLayout from './Layout/AdminLayout';
+
 //Views
 import HomeView from './Views/HomeView';
 import Prueba from './Components/Prueba';
 import DivisionView from './Views/DivisionView';
-
-//Components
-import SiderNav from './Layout/SiderNav';
-import HeaderLayout from './Layout/HeaderLayout';
-
-//AntDesign
-import { Layout } from 'antd';
-const {Content, Footer } = Layout;
+import Login from './Components/Auth/Login';
 
 //Css 
 import '../src/CSS/App.css';
 
-const headerStyle = {
 
-};
 
-const footerStyle = {
-  textAlign: 'center',
-  color: '#fff',
-  backgroundColor: '#4096ff',
-  height:12,
-};
 
 function App() {
 
-  const escuela = "UTM"
   return (
     <>
-        <Layout>
-          <CollapsedProvider>
-            <SiderNav/>
-          </CollapsedProvider>
-          <Layout>
-            <HeaderLayout></HeaderLayout>
-            <Content className='Content'>
-                <Routes>
-                  <Route path='/' element={<DivisionView/>}/>
-                </Routes>
-            </Content>
-            <Footer style={footerStyle}>
-              Hola Footer
-            </Footer>
-          </Layout>
-        </Layout>
+      <Routes>
+        {/* Rutas sin inicio de sesión */}
+        <Route path='/Login' element={ <Login/> }/>
+
+        {/* Rutas con inicio de sesión */}
+        
+        <Route path='/' element={<AdminLayout> <HomeView/> </AdminLayout>}/>
+        <Route path='/Divisiones' element={<AdminLayout> <DivisionView/> </AdminLayout>}/>
+        
+      </Routes>
     </>
   )
 }
