+++ src/CSS/App.css
@@ -1,7 +1,3 @@
 
-.Content{
-    padding-inline: 1rem;
-    background-color: rgb(var(--conifer-50));
-}
