+++ src/CSS/Layout/AdminLayout.css
@@ -1,0 +1,9 @@
+.AdminLayoutAnt{
+    min-height: 100dvh;
+}
+
+
+.AdminLayoutContent{
+    padding-inline: 1rem;
+    background-color: rgb(var(--conifer-50));
+}
