+++ src/CSS/Views/DivisionView.css
@@ -7,3 +7,7 @@
 .SandBoxCol{
     background-color: antiquewhite;
 }
+
+.TittleHomeOptions{
+    font-size: 1.8rem;
+}
