+++ src/Components/Auth/Registro.jsx
