+++ src/Components/General/CardContent.jsx
@@ -1,9 +1,8 @@
-
 //Ant
 import { Card } from 'antd';
 
 //CSS
-import '../CSS/CardContent.css';
+import '../../CSS/General/CardContent.css';
 
 
 export default function CardContent({ children, className }){
