+++ src/index.css
@@ -14,7 +14,6 @@
 
 *{
   box-sizing: border-box;
-  min-width: 0vh;
   padding: 0px;
 }
